Refactor and expand AsyncStorageTable
